requirements.txt
# Pour install, executer: pip install -r requirements.txt

# Librairies cnx source de données
marklogic-python-client==1.3.0
oracledb==3.4.1
psycopg2==2.9.11